# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service responsible for storing and retrieving credentials and presentations",
        "title": "Storage service API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/credentials": {
            "post": {
                "description": "Get credentials from the storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Get credentials from the storage",
                "parameters": [
                    {
                        "description": "Presentation definition details",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/presentation.PresentationDefinition"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "credentials",
                        "schema": {
                            "$ref": "#/definitions/model.GetCredentialModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/credentials/{id}": {
            "put": {
                "description": "Add a credential to the storage",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "Add a credential to the storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The VerifiableCredential raw data to upload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the credential",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/presentations": {
            "post": {
                "description": "Add a presentation to the storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presentations"
                ],
                "summary": "Add a presentation to the storage",
                "parameters": [
                    {
                        "description": "Presentation definition details",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/presentation.PresentationDefinition"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "presentations",
                        "schema": {
                            "$ref": "#/definitions/model.GetCredentialModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/presentations/{id}": {
            "put": {
                "description": "Add a presentation to the storage",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "presentations"
                ],
                "summary": "Add a presentation to the storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The VerifiablePresentation raw data to upload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the presentation",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receipt",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.GetCredentialModel": {
            "type": "object",
            "properties": {
                "credentials": {
                    "type": "object",
                    "additionalProperties": true
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presentation.FilterResult"
                    }
                },
                "receipt": {
                    "type": "string"
                }
            }
        },
        "presentation.Alg": {
            "type": "string",
            "enum": [
                "EdDSA",
                "ES256",
                "PS256"
            ],
            "x-enum-varnames": [
                "EDDSA",
                "ES256",
                "PS256"
            ]
        },
        "presentation.Constraints": {
            "type": "object",
            "properties": {
                "fields": {
                    "description": "The constraints object MAY contain a fields property",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presentation.Field"
                    }
                },
                "limit_disclosure": {
                    "description": "The constraints object MAY contain a limit_disclosure property",
                    "allOf": [
                        {
                            "$ref": "#/definitions/presentation.Disclosure"
                        }
                    ]
                }
            }
        },
        "presentation.Description": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                }
            }
        },
        "presentation.Disclosure": {
            "type": "string",
            "enum": [
                "required",
                "preferred"
            ],
            "x-enum-varnames": [
                "Required",
                "Preferred"
            ]
        },
        "presentation.Field": {
            "type": "object",
            "properties": {
                "filter": {
                    "description": "Optional Field",
                    "allOf": [
                        {
                            "$ref": "#/definitions/presentation.Filter"
                        }
                    ]
                },
                "id": {
                    "description": "Optional Field",
                    "type": "string"
                },
                "name": {
                    "description": "Optional Field",
                    "type": "string"
                },
                "path": {
                    "description": "Mandatory Field",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "purpose": {
                    "description": "Optional Field",
                    "type": "string"
                }
            }
        },
        "presentation.Filter": {
            "type": "object",
            "properties": {
                "pattern": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "presentation.FilterResult": {
            "type": "object",
            "properties": {
                "credentials": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "$ref": "#/definitions/presentation.Description"
                }
            }
        },
        "presentation.Format": {
            "type": "object",
            "properties": {
                "jwt": {
                    "$ref": "#/definitions/presentation.FormatSpecification"
                },
                "jwt_vc": {
                    "$ref": "#/definitions/presentation.FormatSpecification"
                },
                "jwt_vp": {
                    "$ref": "#/definitions/presentation.FormatSpecification"
                },
                "ldp": {
                    "$ref": "#/definitions/presentation.FormatSpecification"
                },
                "ldp_vc": {
                    "$ref": "#/definitions/presentation.FormatSpecification"
                },
                "ldp_vp": {
                    "$ref": "#/definitions/presentation.FormatSpecification"
                },
                "verifiable-credential+sd-jwt": {
                    "$ref": "#/definitions/presentation.FormatSpecification"
                }
            }
        },
        "presentation.FormatSpecification": {
            "type": "object",
            "properties": {
                "alg": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presentation.Alg"
                    }
                },
                "proof_type": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presentation.ProofType"
                    }
                }
            }
        },
        "presentation.InputDescriptor": {
            "type": "object",
            "properties": {
                "constraints": {
                    "$ref": "#/definitions/presentation.Constraints"
                },
                "format": {
                    "$ref": "#/definitions/presentation.Format"
                },
                "group": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                }
            }
        },
        "presentation.PresentationDefinition": {
            "type": "object",
            "properties": {
                "format": {
                    "$ref": "#/definitions/presentation.Format"
                },
                "id": {
                    "type": "string"
                },
                "input_descriptors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presentation.InputDescriptor"
                    }
                },
                "name": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                },
                "submission_requirements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presentation.SubmissionRequirement"
                    }
                }
            }
        },
        "presentation.ProofType": {
            "type": "string",
            "enum": [
                "JsonWebSignature2020",
                "Ed25519Signature2018",
                "EcdsaSecp256k1Signature2019",
                "RsaSignature2018"
            ],
            "x-enum-varnames": [
                "JsonWebSignature2020",
                "Ed25519Signature2018",
                "EcdsaSecp256k1Signature2019",
                "RsaSignature2018"
            ]
        },
        "presentation.Rule": {
            "type": "string",
            "enum": [
                "all",
                "pick"
            ],
            "x-enum-varnames": [
                "All",
                "Pick"
            ]
        },
        "presentation.SubmissionRequirement": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                },
                "rule": {
                    "$ref": "#/definitions/presentation.Rule"
                }
            }
        }
    }
}